update
almost finished the grenade functionality. The mimic is now patrolling and when hit by a grenade, an explosion billboard goes off and a grenade pickup item is meant to spawn in place of the mimic that dissapears. upon pickup, this gives the user another throw of the grenade. lightning bolt pickup is in place